Move functions into their own, separate files & add docs
